Split statisticts tests
