+++ ydb/core/statistics/aggregator/ya.make
@@ -36,3 +36,7 @@ PEERDIR(
 YQL_LAST_ABI_VERSION()
 
 END()
+
+RECURSE_FOR_TESTS(
+    ut
+)
