+++ ydb/core/statistics/database/ya.make
@@ -17,3 +17,7 @@ PEERDIR(
 YQL_LAST_ABI_VERSION()
 
 END()
+
+RECURSE_FOR_TESTS(
+    ut
+)
