+++ ydb/core/statistics/service/ut/ut_basic_statistics.cpp
@@ -1,4 +1,4 @@
-#include "ut_common.h"
+#include <ydb/core/statistics/ut_common/ut_common.h>
 
 #include <ydb/library/actors/testlib/test_runtime.h>
 
@@ -82,7 +82,7 @@ void ValidateRowCount(TTestActorRuntime& runtime, ui32 nodeIndex, TPathId pathId
 
 } // namespace
 
-Y_UNIT_TEST_SUITE(Statistics) {
+Y_UNIT_TEST_SUITE(BasicStatistics) {
 
     Y_UNIT_TEST(Simple) {
         TTestEnv env(1, 1);
