+++ ydb/core/statistics/service/ya.make
@@ -21,3 +21,8 @@ PEERDIR(
 YQL_LAST_ABI_VERSION()
 
 END()
+
+RECURSE_FOR_TESTS(
+    ut
+)
+
