+++ ydb/core/statistics/ut_common/ya.make
@@ -1,0 +1,15 @@
+LIBRARY()
+
+SRCS(
+    ut_common.cpp
+    ut_common.h
+)
+
+PEERDIR(
+    ydb/core/testlib
+)
+
+YQL_LAST_ABI_VERSION()
+
+END()
+
